Add symlink for build.config
The build/build_test.sh script requires the KERNEL_DIR environment
variable to be set, and if it doesn't find that variable in the
build.config file in the repo root, it forces KERNEL_DIR to "common".
This configuration does not work for the HiKey kernels repository.

This patch fixes the issue by creating the symlink:

build.config -> hikey-linaro/build.config

Change-Id: Ib884a2b7c14fc932b01c11eb195256e2c5526d8e
Test: manual build
Signed-off-by: Alessio Balsini <balsini@google.com>