Update patch set 1
Patch Set 1:

I am not against this change. Just please note that the almost-hermetic toolchain has not yet been ported to ACK 4.19. It is only available so far in 5.4 and mainline.

It is probably a good idea to port it to 4.19 first and then pick from there.

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>