Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi, Matthias 
> > Patch Set 1:
> > 
> > I am not against this change. Just please note that the almost-hermetic toolchain has not yet been ported to ACK 4.19. It is only available so far in 5.4 and mainline.
> > 
> > It is probably a good idea to port it to 4.19 first and then pick from there.
> 
> sorry, I am not sure I understand which changes are necessary here,
> could you give me some details about the changes?
> 
> And with the change here:
> https://android-review.googlesource.com/c/kernel/hikey-linaro/+/1259440/1/build.config
> 
> I could build the kernel for hikey960, and it could boot with prebuilt user space images,
> I was wrong somewhere?
> 
> Thanks,
> Yongqin Liu

Hi,

I am not saying you did something wrong (and: nice work! :-)). Just I am saying that android-4.19 is not yet completely switched to this toolchain and things might not be as stable and tested as on 5.4 and mainline. The backport 4.19 was not yet done, so there are no changes yet. But if I do it, then I would do it in android-4.19 and you would either pick it or merge it.

I hope this makes sense?

For now, please go ahead and let me know if things are not working. I am then happy to help.

Cheers,
Matthias

Patch-set: 1
Reviewer: Gerrit User 1402878 <1402878@85c56323-6fa9-3386-8a01-6480fb634889>