Update patch set 2
Patch Set 2: Lint+2

I checked with stubby - the manifest is okay in that regard.
I'm not 100% sure that means the references are all good, but for sure the lint will be negative with a remote like this manifest has. In other words; please give it a try.

Patch-set: 2
Reviewer: Gerrit User 1069528 <1069528@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Lint=+2