Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> since mkdtboimg.py has an actual build description
> 
> https://android.googlesource.com/platform/system/libufdt/+/refs/heads/master/utils/src/Android.bp#49
> 
> I would consider prebuilding it via kernel-build-tools as well. As far as I understand, that would make it a standalone binary with embedded py runtime and therefore independent of any other python runtime.
> 
> What do you think? I would like to avoid adding a lot of additional build tools repositories and glueing them together in every single kernel common manifest.

I tried this in https://android-review.googlesource.com/c/kernel/manifest/+/1331842 but it brings in a lot of dependencies, and also would require running the standalone executable mkdtboimg.py with a custom LD_LIBRARY_PATH. This is going to confuse folks and generate support requests I think when they try and run it directly and get a library error. I think we should just bring in libufdt for this one.

Patch-set: 1
Label: Presubmit-Ready=0