Prepare build of kernel-build-tools with extra STG changes
The is the manifest for master-kernel-build-2022 with the external/stg
hash changed to c54186c03c685557ea57ce8068e4658f63457c20.

This is add to Android 13 the same post-processing of ABI reports seen
in passive ABI monitoring. In particular, the separate labelling of
function and variable symbols is important for ABI change reviewers.

Bug: 248207641
Signed-off-by: Giuliano Procida <gprocida@google.com>
Change-Id: I7de5042757d7c600a9f66120d5cef9c4dafee9ad