Pin trout kernel to 5.10.160_r00
The latest release on the 5.10.x branch causes a boot failure with
trout. Reverting to 5.10.160_r00 allows the system to boot. A root cause
has not been determined, however trout development on 5.10 is
essentially frozen, so we're instead electing to pin the kernel at this
version so we can at least reproduce a working 5.10 configuration.

Change-Id: I33d2b1a6ed83ab49ee0f612322eb142fc0e76874